requirements.txt
# Backend
fastapi
uvicorn
langchain

faiss-cpu
pinecone-client
sentence-transformers
huggingface-hub
pydantic
numpy
scipy
python-dotenv
# Code Generation & Execution
openai
transformers
jupyter
nbconvert
accelerate
# google-colab

# Query Routing (Intent Classification)
spacy
regex
scikit-learn
torch

# Web Interface (GARDO)
# gardo
streamlit

# Additional Utilities
requests
python-dotenv
tqdm

# langchain
langchain
langchain-community
langchain-huggingface

pdfminer.six
unstructured

sentencepiece